+++ asset_test.go
@@ -106,6 +106,17 @@ sub/require.js
 require.js
 `, content)
 
+	content, _ = ReadAsset("/assets/javascripts/require.coffee")
+	assert.Equal(`normal.js
+
+(function() {
+  var a;
+
+  a = 1;
+
+}).call(this);
+`, content)
+
 	content, _ = ReadAsset("/assets/stylesheets/require.css")
 	assert.Equal(`normal.css
 
@@ -115,6 +126,30 @@ sub/require.css
 
 require.css
 `, content)
+
+	content, _ = ReadAsset("/assets/stylesheets/require.scss")
+	assert.Equal(`.normal1-scss {
+  color: blue; }
+sub/normal.css
+
+sub/require.css
+
+.normal1-scss {
+  color: blue; }
+
+.foo .bar {
+  color: red; }`, content)
+
+	// In current can not remove require comment lines
+	content, _ = ReadAsset("/assets/stylesheets/require.sass")
+	assert.Equal(`.normal1-scss {
+  color: blue; }
+sub/normal.css
+
+sub/require.css
+
+.foo {
+  color: red; }`, content)
 
 	_, err = ReadAsset("/assets/javascripts/error.js")
 	assert.Equal(`Asset Not Found: not/found.js