+++ cmd/cmd.go
@@ -1,23 +1,26 @@
 package main
 
 import (
-	"bytes"
 	"flag"
 	"fmt"
 	"github.com/shaoshing/train"
-	"os/exec"
+	. "github.com/shaoshing/train/command"
 )
 
 var helpFlag bool
+var sourcePath string
+var outPath string
 
 func main() {
 	flag.BoolVar(&helpFlag, "h", false, "")
+	flag.StringVar(&sourcePath, "source", "./assets", "")
+	flag.StringVar(&outPath, "out", "./public", "")
 	flag.Parse()
 
 	command := "bundle"
 
 	args := flag.Args()
-	if len(args) == 1 {
+	if len(args) >= 1 {
 		command = args[0]
 	}
 
@@ -28,11 +31,11 @@ func main() {
 
 	switch command {
 	case "bundle":
-		bundle()
+		Bundle(sourcePath, outPath)
 	case "upgrade":
-		upgrade()
+		Upgrade()
 	case "diagnose":
-		diagnose()
+		Diagnose()
 	case "version":
 		fmt.Println("Train version", train.VERSION)
 	case "help":
@@ -42,22 +45,20 @@ func main() {
 	}
 }
 
-func bash(bash string) (out string, err error) {
-	cmd := exec.Command("sh", "-c", bash)
-	var buf bytes.Buffer
-	cmd.Stderr = &buf
-	cmd.Stdout = &buf
-	err = cmd.Run()
-	out = buf.String()
-	return
-}
-
 func showHelp() {
 	fmt.Printf(`usage: train [-h] [command]
 
 OPTIONS
   -h
     Show this help message
+
+  --source
+    Assets source path, default: ./assets
+    example: $ train --source app/assets bundle
+
+  --out
+    Assets output path, default: ./public
+    example: $ train --out /tmp/public bundle
 
 COMMANDS
   bundle [default]