+++ command/upgrade.go
@@ -1,4 +1,4 @@
-package main
+package trainCommand
 
 import (
 	"fmt"
@@ -6,7 +6,7 @@ import (
 
 const CmdBinPath = "$GOPATH/bin/train"
 
-func upgrade() {
+func Upgrade() {
 	bash("go get -u github.com/shaoshing/train")
 	bash("go build -o " + CmdBinPath + " github.com/shaoshing/train/cmd")
 	fmt.Println("Installed latest train command into " + CmdBinPath)